Create ProjectController tests

+++ src/controllers/project.ts
@@ -1,34 +1,11 @@
-import ProjectService from '../services/project'
 import { Request, Response } from 'express'
 import { check, validationResult } from 'express-validator'
 
-import User from '../models/User'
+import ProjectService from '../services/project'
 
 export const validationCreateProject = [
   check('title', 'Title is required').not().isEmpty(),
 ]
-
-const getAllProjects = async (req: Request, res: Response, next: Function) => {
-  try {
-    const projects = await ProjectService.getAllProjects()
-    if (!projects) {
-      res.status(404).json('There are no projects published yet!')
-    }
-    res.json(projects)
-  } catch (error) {
-    res.status(500).json({ error: error })
-  }
-}
-
-const getProjectById = async (req: Request, res: Response, next: Function) => {
-  try {
-    let id = req.params.id || ''
-    const project = await ProjectService.getProjectbyId(id)
-    res.json(project)
-  } catch (error) {
-    res.status(500).json({ error: error })
-  }
-}
 
 const createProject = async (req: Request, res: Response, next: Function) => {
   try {
@@ -37,41 +14,70 @@ const createProject = async (req: Request, res: Response, next: Function) => {
       return res.status(400).json({ errors: errors.array() })
     }
 
-    //TODO create and use user service
-    const user = await User.findById(req.user.id).select('-password')
-
-    if (!user) {
-      return res.status(404).json({ msg: 'User not found' })
-    }
-
     const newProject = {
       title: req.body.title,
-      user: user.id,
+      user: req.user.id,
     }
 
     const createdProject = await ProjectService.createProject(newProject)
     res.status(201).json(createdProject)
   } catch (error) {
-    res.status(500).json({ error: error })
+    res.status(500).json({ errors: [{ msg: 'Server error' }] })
   }
 }
 
-//TODO UpdateProject
+const getProject = async (req: Request, res: Response, next: Function) => {
+  try {
+    let id = req.params.id
+    const project = await ProjectService.getProject(id)
 
-//TODO verify user
+    if (!project) {
+      return res.status(400).json({ errors: [{ msg: 'Project not found' }] })
+    }
+
+    res.status(200).json(project)
+  } catch (error) {
+    res.status(500).json({ errors: [{ msg: 'Server error' }] })
+  }
+}
+
+//TODO updateProject
+
 const deleteProject = async (req: Request, res: Response, next: Function) => {
   try {
     const projectId = req.params.id
-    const deleteResponse = await ProjectService.deleteProject(projectId)
-    res.json(deleteResponse)
+
+    const project = await ProjectService.getProject(projectId)
+
+    if (!project) {
+      return res.status(400).json({ errors: [{ msg: 'Project not found' }] })
+    }
+
+    if (project.user.toString() !== req.user.id) {
+      return res.status(403).json({ errors: [{ msg: 'User not authorized' }] })
+    }
+
+    const deletedProject = await ProjectService.deleteProject(projectId)
+
+    res.status(200).json(deletedProject)
   } catch (error) {
-    res.status(500).json({ error: error })
+    res.status(500).json({ errors: [{ msg: 'Server error' }] })
+  }
+}
+
+const getUserProjects = async (req: Request, res: Response, next: Function) => {
+  try {
+    const projects = await ProjectService.getUserProjects(req.user.id)
+
+    res.status(200).json(projects)
+  } catch (error) {
+    res.status(500).json({ errors: [{ msg: 'Server error' }] })
   }
 }
 
 export default {
-  getAllProjects,
-  getProjectById,
   createProject,
+  getProject,
   deleteProject,
+  getUserProjects,
 }
